fix: move api call to server-side action
